change post id from long to int

+++ toylab-blog/src/main/java/me/chuck/toylab/blog/db/post/PostDAO.java
@@ -21,7 +21,7 @@ public class PostDAO extends AbstractDAO<PostDO> {
     return Optional.ofNullable(persist(postDO));
   }
 
-  public Optional<PostDO> findById(long id) {
+  public Optional<PostDO> findById(int id) {
     return Optional.ofNullable(get(id));
   }
